+++ packages/app/components/PermissionRequest/index.tsx
@@ -1,7 +1,8 @@
-import { ReactElement } from 'react';
+import { ReactElement, useEffect } from 'react';
 
 import { useAppDefinition } from '../AppDefinitionProvider/index.js';
 import { useServiceWorkerRegistration } from '../ServiceWorkerRegistrationProvider/index.js';
+import { useUser } from '../UserProvider/index.js';
 import styles from './index.module.css';
 
 /**
@@ -9,11 +10,20 @@ import styles from './index.module.css';
  */
 export function PermissionRequest(): ReactElement {
   const { definition } = useAppDefinition();
+  const { userInfo } = useUser();
   const { permission, requestPermission, subscribe } = useServiceWorkerRegistration();
 
-  if (definition.notifications && definition.notifications === 'startup') {
+  useEffect(() => {
+    if (definition.notifications === 'opt-in') {
+      return;
+    }
+
+    if (definition.notifications === 'login' && !userInfo?.sub) {
+      return;
+    }
+
     if (window.Notification?.permission === 'denied') {
-      return null;
+      return;
     }
 
     requestPermission().then((p) => {
@@ -21,7 +31,7 @@ export function PermissionRequest(): ReactElement {
         subscribe();
       }
     });
-  }
+  }, [definition.notifications, requestPermission, subscribe, userInfo?.sub]);
 
   return permission === 'pending' ? <div className={`modal-background ${styles.overlay}`} /> : null;
 }