+++ packages/server/utils/actions/index.ts
@@ -6,7 +6,9 @@ import { Mailer } from '../email/Mailer.js';
 import { condition } from './condition.js';
 import { each } from './each.js';
 import { email } from './email.js';
+import { log } from './log.js';
 import { noop } from './noop.js';
+import { notify } from './notify.js';
 import { request } from './request.js';
 import { staticAction } from './static.js';
 import { throwAction } from './throw.js';
@@ -38,15 +40,17 @@ export const actions = {
   link: noop,
   'link.back': noop,
   'link.next': noop,
-  log: noop,
+  log,
   message: noop,
   noop,
+  notify,
   request,
   'resource.create': request,
   'resource.delete': request,
   'resource.get': request,
   'resource.query': request,
   'resource.update': request,
+  'resource.patch': noop,
   'resource.count': request,
   'resource.subscription.status': noop,
   'resource.subscription.subscribe': noop,