+++ packages/web-utils/resource.ts
@@ -1,5 +1,5 @@
 import { mapValues } from '@appsemble/utils';
-import { JsonValue } from 'type-fest';
+import { JsonObject, JsonValue } from 'type-fest';
 
 export function serializeResource(data: any): FormData | JsonValue {
   const assets: Blob[] = [];
@@ -13,8 +13,8 @@ export function serializeResource(data: any): FormData | JsonValue {
     if (value instanceof Date) {
       return value.toJSON();
     }
-    if (value instanceof Object) {
-      return mapValues(value, extractAssets);
+    if (value && typeof value === 'object') {
+      return mapValues(value as JsonObject, extractAssets);
     }
     return value;
   };